+++ robot_descriptions/rby1_description.py
@@ -1,7 +1,10 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
+#
+# SPDX-License-Identifier: Apache-2.0
+# Copyright 2022 Stéphane Caron
 
-"""G1 description."""
+"""YAM URDF description."""
 
 from os import getenv as _getenv
 from os import path as _path
@@ -9,10 +12,10 @@ from os import path as _path
 from ._cache import clone_to_cache as _clone_to_cache
 
 REPOSITORY_PATH: str = _clone_to_cache(
-    "unitree_ros",
+    "rby1_description",
     commit=_getenv("ROBOT_DESCRIPTION_COMMIT", None),
 )
 
-PACKAGE_PATH: str = _path.join(REPOSITORY_PATH, "robots", "g1_description")
+PACKAGE_PATH: str = _path.join(REPOSITORY_PATH, "models/rby1a/urdf")
 
-URDF_PATH: str = _path.join(PACKAGE_PATH, "g1_29dof.urdf")
+URDF_PATH: str = _path.join(PACKAGE_PATH, "model.urdf")